+++ contrib/completions/zoxide.bash
@@ -19,6 +19,9 @@ _zoxide() {
             zoxide,add)
                 cmd="zoxide__subcmd__add"
                 ;;
+            zoxide,dedupe)
+                cmd="zoxide__subcmd__dedupe"
+                ;;
             zoxide,edit)
                 cmd="zoxide__subcmd__edit"
                 ;;
@@ -71,7 +74,7 @@ _zoxide() {
 
     case "${cmd}" in
         zoxide)
-            opts="-h -V --help --version add edit import init query remove"
+            opts="-h -V --help --version add dedupe edit import init query remove"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 1 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0
@@ -99,6 +102,20 @@ _zoxide() {
                     COMPREPLY=($(compgen -f "${cur}"))
                     return 0
                     ;;
+                *)
+                    COMPREPLY=()
+                    ;;
+            esac
+            COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
+            return 0
+            ;;
+        zoxide__subcmd__dedupe)
+            opts="-i -n -h -V --assume-insensitive --dry-run --help --version <pathglob>..."
+            if [[ ${cur} == -* || ${COMP_CWORD} -eq 2 ]] ; then
+                COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
+                return 0
+            fi
+            case "${prev}" in
                 *)
                     COMPREPLY=()
                     ;;